+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.2.0
 
+- added profile classes for categories new in UnifiedPOS 1.16, i.e. DeviceMonitor, GestureControl, GraphicDisplay, IndividualRecognition,
+  SoundPlayer, SoundRecorder, SpeechSynthesis, VideoCapture, VoiceRecognition
 - switched from `Hashtable` to `HashMap` implementation for optimization reasons
 
 ## 3.1.0